Cart officially working
Got cart to finally work, but this required creating a project page in order to do so.
