+++ src/styles/home.css
@@ -28,10 +28,11 @@
     background: #f6f6f6;
     padding-right: 0;
     border-radius: .25rem;
-    padding: 1rem;
+    padding: 1.5rem;
     display: flex;
     justify-content: center;
     align-items: center;
+
 }
 
 .home-title {
@@ -47,6 +48,7 @@
     border-radius: .25rem;
     width: 100%;
     height: auto;
+    border: 4px solid #20150e;
 }
 
 .images {
